Add support to omit storing the source object in the index.
The MappingSettings now have a new property called StoreSettings which
will contain common settings used when storing objects in an index. The
only such setting right now is StoreSource (bool) which indicates
whether or not the source object should be stored in the index. As
before, the default settings will store the source as '$source'.